requirements.txt
Flask
Flask-Login
Flask-FlatPages
Flask-Babel
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Bootstrap-Flask
SQLAlchemy-Utils
Authlib
WTForms-Alchemy
attrs

# simplified markup processors
docutils
textile
markdown2

# pinned
SQLAlchemy<1.3
